python_MINECRAFT/requirements.txt
argparse==1.2.1
distribute==0.6.24
pyFirmata==0.9.5
pyserial==2.7
tailer3==0.3
wsgiref==0.1.2
PyUserInput==0.1.9
pyobjc-framework-Quartz==3.0.4